Add board definition for M5Stack Core2
This is initial support for the device, including most of the basic
peripherals, however there are some unresolved questions and possibly
changes necessary.

The AXP192 power management chip is initialized in `board.c`, and
turns on the 5V boost circuit by default. Without it, any
Stemma/Grove/Qwiic devices connected to port A (and most M5Stack
modules stacked under the board) will not receive power and it will
not be obvious to the user why they aren't working.

However, some M5Stack modules are designed to have their own power
supply and provide 5v power TO the M5Stack rather than draw from it.

Arduino code generally lets users decide whether to turn that boost
circuit on or not, and

Things that work:

* DisplayIO
* Speaker
* Touchscreen (adafruit focaltouch module works)
* WiFi

Things that don't work:

* PDM mic (enabling AUDIOBUSIO_PDMIN results in compiler errors)
* BLE (not enabled for any esp32 device, results in compiler errors)